another bunch of refactoring
