+++ Assets/Scripts/SS3D/Systems/Tile/TileMapCreator/BuildGhost.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a50fdc25aaa4ae4484273f5a646b7e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
